merged

+++ utils/steward-cli/src/commands/info/view_next_index_to_remove.rs
@@ -26,7 +26,10 @@ pub async fn command_view_next_index_to_remove(
 
 fn _print_next_index_to_remove(state_account: &StewardStateAccount) {
     for i in 0..state_account.state.num_pool_validators {
-        let value = state_account.state.validators_to_remove.get_unsafe(i);
+        let value = state_account
+            .state
+            .validators_to_remove
+            .get_unsafe(i as usize);
 
         if value {
             println!("Validator {} is marked for removal", i);
