+++ utils/steward-cli/src/commands/info/view_state.rs
@@ -137,7 +137,7 @@ pub fn format_state(
         "Progress: {:?} / {} ({} remaining)\n",
         state.progress.count(),
         state.num_pool_validators,
-        state.num_pool_validators - state.progress.count()
+        state.num_pool_validators - state.progress.count() as u64
     );
     formatted_string += &format!(
         "Validator Lamport Balances Count: {}\n",
@@ -159,7 +159,7 @@ pub fn format_state(
         "Progress: {:?} / {} ( {} left )\n",
         state.progress.count(),
         state.num_pool_validators,
-        state.num_pool_validators - state.progress.count()
+        state.num_pool_validators - state.progress.count() as u64
     );
     formatted_string += &format!(
         "Start Computing Scores Slot: {}\n",
@@ -205,7 +205,7 @@ fn _print_default_state(
         "Progress: {:?} / {} ({} remaining)\n",
         state.progress.count(),
         state.num_pool_validators,
-        state.num_pool_validators - state.progress.count()
+        state.num_pool_validators - state.progress.count() as u64
     );
     formatted_string += &format!(
         "Validator Lamport Balances Count: {}\n",
@@ -227,7 +227,7 @@ fn _print_default_state(
         "Progress: {:?} / {} ( {} left )\n",
         state.progress.count(),
         state.num_pool_validators,
-        state.num_pool_validators - state.progress.count()
+        state.num_pool_validators - state.progress.count() as u64
     );
     formatted_string += &format!(
         "Start Computing Scores Slot: {}\n",
